+++ src/CoreMetadata/reduxer.js
@@ -3,7 +3,8 @@ import CoreMetadataHeader from './CoreMetadataHeader';
 import FileTypePicture from '../components/FileTypePicture';
 import CoreMetadataTable from './CoreMetadataTable';
 import { coreMetadataPath, userapiPath } from '../localconf';
-import { fetchWithCreds, updatePopup } from '../actions';
+import { updatePopup } from '../actions';
+import { fetchWithCreds } from '../actions.thunk';
 
 export const generateSignedURL = (objectId) => (dispatch) =>
   fetchWithCreds({
